Rename temp schema files
