+++ packages/lesswrong/components/events/EventsHome.tsx
@@ -18,7 +18,7 @@ import { getBrowserLocalStorage } from '../editor/localStorageHandlers';
 import Geosuggest from 'react-geosuggest';
 import Button from '@material-ui/core/Button';
 import { forumTypeSetting } from '../../lib/instanceSettings';
-import { EVENT_TYPES } from '../../lib/collections/posts/temp_schema';
+import { EVENT_TYPES } from '../../lib/collections/posts/temp_schema_2';
 import Input from '@material-ui/core/Input';
 import OutlinedInput from '@material-ui/core/OutlinedInput';
 import Checkbox from '@material-ui/core/Checkbox';
