+++ packages/lesswrong/components/localGroups/EventNotificationsDialog.tsx
@@ -12,7 +12,7 @@ import InputAdornment from '@material-ui/core/InputAdornment';
 import FormLabel from '@material-ui/core/FormLabel';
 import Checkbox from '@material-ui/core/Checkbox';
 import { geoSuggestStyles, useGoogleMaps } from '../form-components/LocationFormComponent'
-import { MAX_NOTIFICATION_RADIUS } from '../../lib/collections/users/temp_schema'
+import { MAX_NOTIFICATION_RADIUS } from '../../lib/collections/users/temp_schema_2'
 import { forumTypeSetting } from '../../lib/instanceSettings';
 import deepmerge from 'deepmerge';
