+++ packages/lesswrong/components/tagging/WikiGradeDisplay.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { registerComponent, Components } from '../../lib/vulcan-lib';
-import { wikiGradeDefinitions } from '../../lib/collections/tags/temp_schema';
+import { wikiGradeDefinitions } from '../../lib/collections/tags/temp_schema_2';
 import StarIcon from '@material-ui/icons/Star';
 import { Link } from '../../lib/reactRouterWrapper';
 import { forumTypeSetting, siteNameWithArticleSetting, taggingNameIsSet, taggingNamePluralSetting, taggingNameSetting } from '../../lib/instanceSettings';
