+++ packages/lesswrong/lib/collections/comments/index.ts
@@ -1,6 +1,6 @@
 export * from './collection';
 
-import './temp_schema';
+import './temp_schema_2';
 import './helpers';
 import './permissions';
 import './views';
