+++ packages/lesswrong/lib/collections/posts/temp_schema_2.ts
