+++ packages/lesswrong/lib/collections/users/collection.ts
@@ -1,4 +1,4 @@
-import schema from './temp_schema';
+import schema from './temp_schema_2';
 import { createCollection, addGraphQLQuery, addGraphQLResolvers } from '../../vulcan-lib';
 import { userOwns, userCanDo } from '../../vulcan-users/permissions';
 import { addUniversalFields, getDefaultMutations, getDefaultResolvers } from '../../collectionUtils';
